[Improve]: sql auto complete improve
