+++ streamx-console/streamx-console-service/pom.xml
@@ -668,9 +668,6 @@
                     </execution>
                 </executions>
             </plugin>
-
-
-
         </plugins>
     </build>
