+++ streamx-console/streamx-console-service/src/main/resources/sql-rev.dict
